@@ -16,7 +16,7 @@ version: 1.0.65+65
 
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.11.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,46 +26,39 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   amplitude_flutter: ^3.2.1
-
   clipboard: ^0.1.3
   cached_network_image: ^3.0.0
   font_awesome_flutter:
-  flutter_ffmpeg: ^0.2.10
+  flutter_ffmpeg:
   in_app_review: ^2.0.3
   loading_animations: "^2.1.0"
   loading_indicator: ^3.0.1
   audioplayer: ^0.8.1
-  dio: ^4.0.0
+  dio: ^5.1.1
   rflutter_alert: ^2.0.2
-  percent_indicator: ^3.0.1
-  dismissible_page: ^0.6.4
-  #google_sign_in: ^4.4.1
-#  firebase_core: ^0.4.4
+  percent_indicator: ^4.2.3
+  dismissible_page: ^1.0.2
   linkable: ^3.0.1
-  firebase_analytics: ^8.1.0
-#  color_thief_flutter: ^1.0.2
+  firebase_analytics: ^10.2.1
   modal_progress_hud: ^0.1.3
   datetime_picker_formfield: ^2.0.0
   provider: ^6.0.0
-  file_picker: ^4.0.0
-  autocomplete_textfield: ^1.7.3
+  file_picker: ^5.2.10
+  autocomplete_textfield: ^2.0.1
   flutter_datetime_picker: ^1.3.8
   shimmer: ^2.0.0
   shared_preferences: ^2.0.7
-  firebase_messaging: ^10.0.0
-#  flutter_share_file: ^1.2.1+6
+  firebase_messaging: ^14.4.1
   flutter_swiper: ^1.1.6
   palette_generator: ^0.3.2
-  modal_bottom_sheet: ^2.0.0
+#  modal_bottom_sheet: ^2.0.0
   progress_dialog: ^1.2.4
   carousel_slider: ^4.0.0
   flutter_share: ^2.0.0
   transparent_image: ^2.0.0
-#  video_player: ^1.0.1
   autocomplete_textfield_modified: ^1.6.11
   marquee: ^2.0.0
   timeago: ^3.0.2
-#  in_app_update: ^2.0.0
   flutter_spinkit: ^5.0.0
   assets_audio_player: ^3.0.3+6
   koukicons: ^2.0.3
@@ -78,68 +71,50 @@ dependencies:
   scrollable_positioned_list: any
   animated_text_kit: ^4.2.1
   double_back_to_close_app: ^2.0.1
-#  in_app_update: any
-#  flutter_image_compress: ^1.0.0
-  webview_flutter: ^2.3.0
+  webview_flutter: ^4.2.0
   url_launcher: ^6.0.3
-  flutter_html: ^2.1.1
-  image: ^3.0.2
-  intl_phone_field: ^2.0.0
-#  advance_pdf_viewer: ^1.2.1+2
-#  flutter_media_notification: ^1.2.6
-#  flutter_inappwebview: ^5.3.2
+  flutter_html: ^3.0.0-alpha.6
+  image: ^4.0.15
+  intl_phone_field: ^3.1.0
   otp_text_field: ^1.1.0
   pin_code_fields: ^7.0.0
   async: ^2.4.2
-#  flutter_typeahead: ^2.0.0
   fluttertoast: ^8.0.3
   color_thief_flutter: ^1.0.2
   sqflite: ^2.0.0+3
   flutter_local_notifications: ^8.1.1+1
   dotted_border: ^2.0.0-nullsafety.0
-  intl: ^0.17.0
-  permission_handler: ^8.1.4+2
+  intl: ^0.18.1
+  permission_handler: ^10.2.0
   flutter_inappwebview: ^5.3.2
   add_2_calendar: ^2.1.1
- # flutter_svg: ^0.19.0
- # firebase_auth: ^0.16.1
- # flutter_signin_button: ^1.1.0
-#  just_audio: ^0.8.0
-  firebase_auth: ^3.1.0
-  firebase_core: ^1.2.0
+  firebase_auth: ^4.2.2
+  firebase_core: ^2.10.0
   flutter_signin_button: ^2.0.0
   flutter_downloader: ^1.5.2
   form_field_validator: ^1.0.1
-  google_fonts: ^2.0.0
-  google_sign_in: ^5.0.7
+  google_fonts: ^4.0.3
+  google_sign_in: ^6.1.0
   hive: ^2.0.4
   http: ^0.13.0
-  lottie: ^1.0.1
+  lottie: ^2.3.2
   expandable: ^5.0.1
-  sign_in_with_apple: ^3.0.0
+  sign_in_with_apple: ^4.3.0
   page_transition: ^2.0.2
   path_provider: any
   uni_links: ^0.5.0
-  badges: ^2.0.1
+  badges: ^3.1.1
   jitsi_meet: ^4.0.0
-#  slider_controller: ^0.0.3
   flutter_vibrate: ^1.3.0
   mime: ^1.0.1
   chewie: ^1.3.0
   video_player: ^2.3.0
-#  better_player: any
   flutter_sticky_header: ^0.6.1
-#  pip_view:
-#  flutter_riverpod: ^0.13.1+1
   miniplayer: ^1.0.1
 
 
-
-
-
-
 dev_dependencies:
-  flutter_launcher_icons: ^0.9.0
+  flutter_launcher_icons: ^0.12.0
   flutter_test:
     sdk: flutter
 #  local_auth: ^0.6.3+4
