@@ -10,9 +10,10 @@ import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-// import 'package:flutter_inappwebview/flutter_inappwebview.dart';
+import 'package:flutter_inappwebview/flutter_inappwebview.dart';
+
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:webview_flutter/webview_flutter.dart';
+
 
 class HiveAccount extends StatefulWidget {
   static const String id = "Hive Signer";
@@ -199,50 +200,49 @@ class _HiveAccountState extends State<HiveAccount> {
         children: [
           SizedBox.expand(
             child: Center(
-              child: WebView(
-                gestureRecognizers: Set()
-                  ..add(
-                    Factory<VerticalDragGestureRecognizer>(
-                      () => VerticalDragGestureRecognizer(),
-                    ), // or null
-                  ),
-                javascriptMode: JavascriptMode.unrestricted,
-                initialUrl:
-                    'https://hivesigner.com/oauth2/authorize?client_id=aureal&redirect_uri=%3Dhttp%253A%252F%252Flocalhost%253A3000%26&response_type=code&scope=offline,comment,vote,comment_option,custom_json',
-                onPageStarted: (url) async {
-                  SharedPreferences prefs =
-                      await SharedPreferences.getInstance();
-
-                  var uri = Uri.parse(url.toString());
-                  uri.queryParameters.forEach((key, value) {
-                    if (key == 'access_token' ||
-                        key == 'username' ||
-                        key == 'code') {
-                      if (key == 'username' &&
-                          prefs.getString('access_token') != null &&
-                          prefs.getString('code') != null) {
-                        prefs.setString('HiveUserName', value);
-
-                        print(prefs.getString('code'));
-
-// registerHiveUser();
-
-                      } else {
-                        prefs.setString('$key', value);
-                      }
-
-                      print(prefs.getString(key));
-                      if (prefs.getString('code') != null) {
-                        print(
-                            'this is the code motherfucker: ${prefs.getString('code')}');
-                      }
-                      hiveAuth();
-                    }
-                  });
-                },
-              ),
-            ),
-          ),
+//               child: InAppWebView(
+//                 gestureRecognizers: Set()
+//                   ..add(
+//                     Factory<VerticalDragGestureRecognizer>(
+//                       () => VerticalDragGestureRecognizer(),
+//                     ), // or null
+//                   ),
+//                 initialUrl: URLRequest(
+//                     'https://hivesigner.com/oauth2/authorize?client_id=aureal&redirect_uri=%3Dhttp%253A%252F%252Flocalhost%253A3000%26&response_type=code&scope=offline,comment,vote,comment_option,custom_json',
+//                 onPageStarted: (url) async {
+//                   SharedPreferences prefs =
+//                       await SharedPreferences.getInstance();
+//
+//                   var uri = Uri.parse(url.toString());
+//                   uri.queryParameters.forEach((key, value) {
+//                     if (key == 'access_token' ||
+//                         key == 'username' ||
+//                         key == 'code') {
+//                       if (key == 'username' &&
+//                           prefs.getString('access_token') != null &&
+//                           prefs.getString('code') != null) {
+//                         prefs.setString('HiveUserName', value);
+//
+//                         print(prefs.getString('code'));
+//
+// // registerHiveUser();
+//
+//                       } else {
+//                         prefs.setString('$key', value);
+//                       }
+//
+//                       print(prefs.getString(key));
+//                       if (prefs.getString('code') != null) {
+//                         print(
+//                             'this is the code motherfucker: ${prefs.getString('code')}');
+//                       }
+//                       hiveAuth();
+//                     }
+//                   });
+//                 },
+//               ),
+//             ),
+          ),)
         ],
       ),
     );
