@@ -10,7 +10,7 @@ import 'package:flutter/material.dart';
 // import 'package:flutter_twitter/flutter_twitter.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
+
 import 'package:provider/provider.dart';
 
 import 'Auth.dart';
@@ -290,7 +290,7 @@ class _WelcomeState extends State<Welcome> {
                   InkWell(
                     onTap: () {
                       print("Hive Signer Activated");
-                      // showBarModalBottomSheet(
+                      // showBottomSheet(
                       //     context: context,
                       //     builder: (context) {
                       //       return HiveAccount();
