@@ -28,7 +28,7 @@ import 'package:flutter_share/flutter_share.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:http/http.dart' as http;
-import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
+
 import 'package:modal_progress_hud/modal_progress_hud.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -369,11 +369,11 @@ class _ProfileState extends State<Profile> with SingleTickerProviderStateMixin {
     //   super.dispose();
     // }
 
-    Future<void> _pullRefresh() async {
-      print('proceedd');
-
-      await (getUserDetails());
-    }
+    // Future<void> _pullRefresh() async {
+    //   print('proceedd');
+    //
+    //   await (getUserDetails());
+    // }
 
     Future<bool> _onBackPressed() async {
       Navigator.pushNamedAndRemoveUntil(
@@ -499,7 +499,7 @@ class _ProfileState extends State<Profile> with SingleTickerProviderStateMixin {
                                                         },
                                                         child: InkWell(
                                                           onTap: () {
-                                                            showBarModalBottomSheet(
+                                                            showBottomSheet(
                                                                 context:
                                                                     context,
                                                                 builder:
@@ -627,713 +627,710 @@ class _ProfileState extends State<Profile> with SingleTickerProviderStateMixin {
                             ),
                           ];
                         },
-                        body: RefreshIndicator(
-                          onRefresh: _pullRefresh,
-                          child: ListView(children: [
-                            Padding(
-                              padding: const EdgeInsets.all(8.0),
-                              child: Column(
-                                crossAxisAlignment: CrossAxisAlignment.start,
-                                mainAxisSize: MainAxisSize.min,
-                                children: [
-                                  ListTile(
-                                    title: Text(
-                                      "Your Podcasts",
-                                      textScaleFactor: 1.0,
-                                      style: TextStyle(
-                                          fontSize:
-                                              SizeConfig.safeBlockHorizontal *
-                                                  4),
-                                    ),
+                        body: ListView(children: [
+                          Padding(
+                            padding: const EdgeInsets.all(8.0),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              mainAxisSize: MainAxisSize.min,
+                              children: [
+                                ListTile(
+                                  title: Text(
+                                    "Your Podcasts",
+                                    textScaleFactor: 1.0,
+                                    style: TextStyle(
+                                        fontSize:
+                                            SizeConfig.safeBlockHorizontal *
+                                                4),
                                   ),
-                                  Container(
-                                    height: MediaQuery.of(context).size.height /
-                                        6.2,
-                                    child: ListView(
-                                      scrollDirection: Axis.horizontal,
-                                      children: [
-                                        for (var v in podcastList)
-                                          Padding(
-                                            padding:
-                                                const EdgeInsets.only(left: 5),
-                                            child: InkWell(
-                                              onTap: () {
-                                                Navigator.push(context,
-                                                    CupertinoPageRoute(
-                                                        builder: (context) {
-                                                  return PodcastView(v['id']);
-                                                }));
+                                ),
+                                Container(
+                                  height: MediaQuery.of(context).size.height /
+                                      6.2,
+                                  child: ListView(
+                                    scrollDirection: Axis.horizontal,
+                                    children: [
+                                      for (var v in podcastList)
+                                        Padding(
+                                          padding:
+                                              const EdgeInsets.only(left: 5),
+                                          child: InkWell(
+                                            onTap: () {
+                                              Navigator.push(context,
+                                                  CupertinoPageRoute(
+                                                      builder: (context) {
+                                                return PodcastView(v['id']);
+                                              }));
+                                            },
+                                            child: Container(
+                                              width: MediaQuery.of(context)
+                                                      .size
+                                                      .width /
+                                                  4.5,
+                                              child: Column(
+                                                children: [
+                                                  Container(
+                                                    decoration: BoxDecoration(
+                                                      boxShadow: [
+                                                        new BoxShadow(
+                                                          color: Colors
+                                                              .black54
+                                                              .withOpacity(
+                                                                  0.2),
+                                                          blurRadius: 10.0,
+                                                        ),
+                                                      ],
+                                                      color: themeProvider
+                                                                  .isLightTheme ==
+                                                              true
+                                                          ? Colors.white
+                                                          : Color(0xff1a1a1a),
+                                                      borderRadius:
+                                                          BorderRadius
+                                                              .circular(15),
+                                                    ),
+                                                    width:
+                                                        MediaQuery.of(context)
+                                                                .size
+                                                                .width /
+                                                            4.5,
+                                                    height:
+                                                        MediaQuery.of(context)
+                                                                .size
+                                                                .width /
+                                                            4.5,
+                                                    child: CachedNetworkImage(
+                                                      imageUrl: v['image'],
+                                                      imageBuilder: (context,
+                                                          imageProvider) {
+                                                        return Container(
+                                                          height: MediaQuery.of(
+                                                                      context)
+                                                                  .size
+                                                                  .width /
+                                                              4.5,
+                                                          width: MediaQuery.of(
+                                                                      context)
+                                                                  .size
+                                                                  .width /
+                                                              4.5,
+                                                          decoration: BoxDecoration(
+                                                              borderRadius:
+                                                                  BorderRadius
+                                                                      .circular(
+                                                                          10),
+                                                              image: DecorationImage(
+                                                                  image:
+                                                                      imageProvider,
+                                                                  fit: BoxFit
+                                                                      .cover)),
+                                                        );
+                                                      },
+                                                    ),
+                                                  ),
+                                                  Padding(
+                                                    padding: const EdgeInsets
+                                                            .symmetric(
+                                                        vertical: 8),
+                                                    child:
+                                                        Text("${v['name']}"),
+                                                  )
+                                                ],
+                                              ),
+                                            ),
+                                          ),
+                                        ),
+
+                                      Padding(
+                                        padding:
+                                            const EdgeInsets.only(left: 5),
+                                        child: Column(
+                                          crossAxisAlignment:
+                                              CrossAxisAlignment.start,
+                                          children: <Widget>[
+                                            GestureDetector(
+                                              onTap: () async {
+                                                if (prefs.getString(
+                                                        'HiveUserName') ==
+                                                    null) {
+                                                  showBottomSheet(
+                                                      context: context,
+                                                      builder: (context) {
+                                                        return HiveDetails();
+                                                      });
+                                                } else {
+                                                  showBottomSheet(
+                                                      context: context,
+                                                      builder: (context) {
+                                                        return EmailVerificationDialog(
+                                                          username:
+                                                              prefs.getString(
+                                                                  'userName'),
+                                                        );
+                                                      });
+                                                }
                                               },
                                               child: Container(
+                                                decoration: BoxDecoration(
+                                                  boxShadow: [
+                                                    new BoxShadow(
+                                                      color: Colors.black54
+                                                          .withOpacity(0.2),
+                                                      blurRadius: 10.0,
+                                                    ),
+                                                  ],
+                                                  color: themeProvider
+                                                              .isLightTheme ==
+                                                          true
+                                                      ? Colors.white
+                                                      : Color(0xff1a1a1a),
+                                                  borderRadius:
+                                                      BorderRadius.circular(
+                                                          15),
+                                                ),
                                                 width: MediaQuery.of(context)
                                                         .size
                                                         .width /
                                                     4.5,
-                                                child: Column(
-                                                  children: [
-                                                    Container(
-                                                      decoration: BoxDecoration(
-                                                        boxShadow: [
-                                                          new BoxShadow(
-                                                            color: Colors
-                                                                .black54
-                                                                .withOpacity(
-                                                                    0.2),
-                                                            blurRadius: 10.0,
-                                                          ),
-                                                        ],
-                                                        color: themeProvider
-                                                                    .isLightTheme ==
-                                                                true
-                                                            ? Colors.white
-                                                            : Color(0xff1a1a1a),
-                                                        borderRadius:
-                                                            BorderRadius
-                                                                .circular(15),
+                                                height: MediaQuery.of(context)
+                                                        .size
+                                                        .width /
+                                                    4.5,
+                                                child: Center(
+                                                  child: Column(
+                                                    mainAxisAlignment:
+                                                        MainAxisAlignment
+                                                            .center,
+                                                    children: [
+                                                      Icon(
+                                                        Icons.add,
                                                       ),
-                                                      width:
-                                                          MediaQuery.of(context)
-                                                                  .size
-                                                                  .width /
-                                                              4.5,
-                                                      height:
-                                                          MediaQuery.of(context)
-                                                                  .size
-                                                                  .width /
-                                                              4.5,
-                                                      child: CachedNetworkImage(
-                                                        imageUrl: v['image'],
-                                                        imageBuilder: (context,
-                                                            imageProvider) {
-                                                          return Container(
-                                                            height: MediaQuery.of(
-                                                                        context)
-                                                                    .size
-                                                                    .width /
-                                                                4.5,
-                                                            width: MediaQuery.of(
-                                                                        context)
-                                                                    .size
-                                                                    .width /
-                                                                4.5,
-                                                            decoration: BoxDecoration(
-                                                                borderRadius:
-                                                                    BorderRadius
-                                                                        .circular(
-                                                                            10),
-                                                                image: DecorationImage(
-                                                                    image:
-                                                                        imageProvider,
-                                                                    fit: BoxFit
-                                                                        .cover)),
-                                                          );
-                                                        },
-                                                      ),
-                                                    ),
-                                                    Padding(
-                                                      padding: const EdgeInsets
-                                                              .symmetric(
-                                                          vertical: 8),
-                                                      child:
-                                                          Text("${v['name']}"),
-                                                    )
-                                                  ],
+                                                    ],
+                                                  ),
                                                 ),
                                               ),
                                             ),
-                                          ),
-
-                                        Padding(
-                                          padding:
-                                              const EdgeInsets.only(left: 5),
-                                          child: Column(
-                                            crossAxisAlignment:
-                                                CrossAxisAlignment.start,
-                                            children: <Widget>[
-                                              GestureDetector(
-                                                onTap: () async {
-                                                  if (prefs.getString(
-                                                          'HiveUserName') ==
-                                                      null) {
-                                                    showBarModalBottomSheet(
-                                                        context: context,
-                                                        builder: (context) {
-                                                          return HiveDetails();
-                                                        });
-                                                  } else {
-                                                    showBarModalBottomSheet(
-                                                        context: context,
-                                                        builder: (context) {
-                                                          return EmailVerificationDialog(
-                                                            username:
-                                                                prefs.getString(
-                                                                    'userName'),
-                                                          );
-                                                        });
-                                                  }
-                                                },
-                                                child: Container(
-                                                  decoration: BoxDecoration(
-                                                    boxShadow: [
-                                                      new BoxShadow(
-                                                        color: Colors.black54
-                                                            .withOpacity(0.2),
-                                                        blurRadius: 10.0,
-                                                      ),
-                                                    ],
-                                                    color: themeProvider
-                                                                .isLightTheme ==
-                                                            true
-                                                        ? Colors.white
-                                                        : Color(0xff1a1a1a),
-                                                    borderRadius:
-                                                        BorderRadius.circular(
-                                                            15),
-                                                  ),
-                                                  width: MediaQuery.of(context)
-                                                          .size
-                                                          .width /
-                                                      4.5,
-                                                  height: MediaQuery.of(context)
-                                                          .size
-                                                          .width /
-                                                      4.5,
-                                                  child: Center(
-                                                    child: Column(
-                                                      mainAxisAlignment:
-                                                          MainAxisAlignment
-                                                              .center,
-                                                      children: [
-                                                        Icon(
-                                                          Icons.add,
-                                                        ),
-                                                      ],
-                                                    ),
-                                                  ),
-                                                ),
-                                              ),
-                                              Padding(
-                                                padding:
-                                                    const EdgeInsets.symmetric(
-                                                        vertical: 8,
-                                                        horizontal: 5),
-                                                child: Text("Add a podcast"),
-                                              ),
-                                            ],
-                                          ),
+                                            Padding(
+                                              padding:
+                                                  const EdgeInsets.symmetric(
+                                                      vertical: 8,
+                                                      horizontal: 5),
+                                              child: Text("Add a podcast"),
+                                            ),
+                                          ],
                                         ),
-                                        // Padding(
-                                        //   padding:
-                                        //       const EdgeInsets.fromLTRB(15, 8, 0, 8),
-                                        //   child: Column(
-                                        //     children: [
-                                        //
-                                        //       Container(
-                                        //         decoration: BoxDecoration(
-                                        //             color: Color(0xff1a1a1a),
-                                        //             borderRadius:
-                                        //                 BorderRadius.circular(10)),
-                                        //         width:
-                                        //             MediaQuery.of(context).size.width /
-                                        //                 4.5,
-                                        //         height:
-                                        //             MediaQuery.of(context).size.width /
-                                        //                 4.5,
-                                        //         child: Icon(Icons.add),
-                                        //       ),
-                                        //       Padding(
-                                        //         padding: const EdgeInsets.all(8.0),
-                                        //         child: Text("Add a podcast"),
-                                        //       )
-                                        //     ],
-                                        //   ),
-                                        // ),
-                                      ],
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                            // Container(
-                            //   height: MediaQuery.of(context).size.height * 0.26,
-                            //   child: Column(
-                            //     crossAxisAlignment: CrossAxisAlignment.start,
-                            //     mainAxisSize: MainAxisSize.min,
-                            //     children: [
-                            //       Padding(
-                            //         padding: const EdgeInsets.symmetric(
-                            //             horizontal: 15, vertical: 10),
-                            //         child: Text(
-                            //           "Your Live Rooms",
-                            //           textScaleFactor: 1.0,
-                            //           style: TextStyle(
-                            //               fontSize: SizeConfig.safeBlockHorizontal * 5),
-                            //         ),
-                            //       ),
-                            //       Container(
-                            //         height: MediaQuery.of(context).size.height * 0.2,
-                            //         child: ListView(
-                            //           scrollDirection: Axis.horizontal,
-                            //           children: [
-                            //             Padding(
-                            //               padding:
-                            //                   const EdgeInsets.fromLTRB(15, 8, 0, 8),
-                            //               child: Column(
-                            //                 children: [
-                            //                   Container(
-                            //                     decoration: BoxDecoration(
-                            //                         color: themeProvider
-                            //                             .isLightTheme ==
-                            //                             true
-                            //                             ? Color(0xffE8E8E8)
-                            //                             : Color(0xff1a1a1a),
-                            //                         borderRadius:
-                            //                             BorderRadius.circular(10)),
-                            //                     width:
-                            //                         MediaQuery.of(context).size.width /
-                            //                             4.5,
-                            //                     height:
-                            //                         MediaQuery.of(context).size.width /
-                            //                             4.5,
-                            //                     child: Icon(Icons.add),
-                            //                   ),
-                            //                   Padding(
-                            //                     padding: const EdgeInsets.all(8.0),
-                            //                     child: Text("Add a podcast"),
-                            //                   )
-                            //                 ],
-                            //               ),
-                            //             ),
-                            //             Padding(
-                            //               padding:
-                            //                   const EdgeInsets.fromLTRB(15, 8, 0, 8),
-                            //               child: Column(
-                            //                 children: [
-                            //                   Container(
-                            //                     decoration: BoxDecoration(
-                            //                         color: Color(0xff1a1a1a),
-                            //                         borderRadius:
-                            //                             BorderRadius.circular(10)),
-                            //                     width:
-                            //                         MediaQuery.of(context).size.width /
-                            //                             4.5,
-                            //                     height:
-                            //                         MediaQuery.of(context).size.width /
-                            //                             4.5,
-                            //                     child: Icon(Icons.add),
-                            //                   ),
-                            //                   Padding(
-                            //                     padding: const EdgeInsets.all(8.0),
-                            //                     child: Text("Add a podcast"),
-                            //                   )
-                            //                 ],
-                            //               ),
-                            //             ),
-                            //             Padding(
-                            //               padding:
-                            //                   const EdgeInsets.fromLTRB(15, 8, 0, 8),
-                            //               child: Column(
-                            //                 children: [
-                            //                   Container(
-                            //                     decoration: BoxDecoration(
-                            //                         color: Color(0xff1a1a1a),
-                            //                         borderRadius:
-                            //                             BorderRadius.circular(10)),
-                            //                     width:
-                            //                         MediaQuery.of(context).size.width /
-                            //                             4.5,
-                            //                     height:
-                            //                         MediaQuery.of(context).size.width /
-                            //                             4.5,
-                            //                     child: Icon(Icons.add),
-                            //                   ),
-                            //                   Padding(
-                            //                     padding: const EdgeInsets.all(8.0),
-                            //                     child: Text("Add a podcast"),
-                            //                   )
-                            //                 ],
-                            //               ),
-                            //             ),
-                            //             Padding(
-                            //               padding:
-                            //                   const EdgeInsets.fromLTRB(15, 8, 0, 8),
-                            //               child: Column(
-                            //                 children: [
-                            //                   Container(
-                            //                     decoration: BoxDecoration(
-                            //                         color: themeProvider
-                            //                             .isLightTheme ==
-                            //                             true
-                            //                             ? Color(0xffE8E8E8)
-                            //                             : Color(0xff1a1a1a),
-                            //                         borderRadius:
-                            //                             BorderRadius.circular(10)),
-                            //                     width:
-                            //                         MediaQuery.of(context).size.width /
-                            //                             4.5,
-                            //                     height:
-                            //                         MediaQuery.of(context).size.width /
-                            //                             4.5,
-                            //                     child: Icon(Icons.add),
-                            //                   ),
-                            //                   Padding(
-                            //                     padding: const EdgeInsets.all(8.0),
-                            //                     child: Text("Add a podcast"),
-                            //                   )
-                            //                 ],
-                            //               ),
-                            //             ),
-                            //           ],
-                            //         ),
-                            //       ),
-                            //     ],
-                            //   ),
-                            // ),
-                            ListTile(
-                              onTap: () {
-                                Navigator.push(
-                                    context,
-                                    CupertinoPageRoute(
-                                        builder: (context) =>
-                                            ReferralProgram()));
-                              },
-                              title: Text(
-                                "Invite ",
-                                textScaleFactor: mediaQueryData.textScaleFactor
-                                    .clamp(0.5, 1.5)
-                                    .toDouble(),
-                                style: TextStyle(
-                                    //  color: Colors.white,
-                                    fontWeight: FontWeight.w700,
-                                    fontSize:
-                                        SizeConfig.safeBlockHorizontal * 4),
-                              ),
-                              trailing: Icon(Icons.arrow_forward_ios_rounded,
-                                  size: 15),
-                              subtitle: Text(
-                                "Invite friends and earn rewards",
-                                textScaleFactor: mediaQueryData.textScaleFactor
-                                    .clamp(0.5, 0.8)
-                                    .toDouble(),
-                                style: TextStyle(
-                                    //       color: Colors.white70,
-                                    fontWeight: FontWeight.w300,
-                                    fontSize:
-                                        SizeConfig.safeBlockHorizontal * 3.4),
-                              ),
-                            ),
-                            ListTile(
-                              onTap: () {
-                                Navigator.push(
-                                    context,
-                                    CupertinoPageRoute(
-                                        builder: (context) => Rewards()));
-                              },
-                              title: Text(
-                                "Your rewards",
-                                textScaleFactor: mediaQueryData.textScaleFactor
-                                    .clamp(0.5, 1.5)
-                                    .toDouble(),
-                                style: TextStyle(
-                                    //  color: Colors.white,
-                                    fontWeight: FontWeight.w700,
-                                    fontSize:
-                                        SizeConfig.safeBlockHorizontal * 4),
-                              ),
-                              trailing: Icon(Icons.arrow_forward_ios_rounded,
-                                  size: 15),
-                              subtitle: Text(
-                                "Check your rewards",
-                                textScaleFactor: mediaQueryData.textScaleFactor
-                                    .clamp(0.5, 0.8)
-                                    .toDouble(),
-                                style: TextStyle(
-                                    //       color: Colors.white70,
-                                    fontWeight: FontWeight.w300,
-                                    fontSize:
-                                        SizeConfig.safeBlockHorizontal * 3.4),
-                              ),
-                            ),
-                            ListTile(
-                              onTap: () {
-                                Navigator.push(
-                                    context,
-                                    CupertinoPageRoute(
-                                        builder: (context) => PublicProfile(
-                                            userId:
-                                                prefs.getString("userId"))));
-                              },
-                              title: Text(
-                                "Public Profile",
-                                textScaleFactor: mediaQueryData.textScaleFactor
-                                    .clamp(0.5, 1.5)
-                                    .toDouble(),
-                                style: TextStyle(
-                                    //  color: Colors.white,
-                                    fontWeight: FontWeight.w700,
-                                    fontSize:
-                                        SizeConfig.safeBlockHorizontal * 4),
-                              ),
-                              trailing: Icon(Icons.arrow_forward_ios_rounded,
-                                  size: 15),
-                              subtitle: Text(
-                                "See how others see you",
-                                textScaleFactor: mediaQueryData.textScaleFactor
-                                    .clamp(0.5, 0.8)
-                                    .toDouble(),
-                                style: TextStyle(
-                                    //       color: Colors.white70,
-                                    fontWeight: FontWeight.w300,
-                                    fontSize:
-                                        SizeConfig.safeBlockHorizontal * 3.4),
-                              ),
-                            ),
-
-                            // Padding(
-                            //   padding: const EdgeInsets.all(15.0),
-                            //   child: InkWell(
-                            //     onTap: () {
-                            //       Navigator.push(
-                            //           context,
-                            //           CupertinoPageRoute(
-                            //               builder: (context) =>
-                            //                   ReferralProgram()));
-                            //     },
-                            //     child: Container(
-                            //       child: Row(
-                            //         mainAxisAlignment:
-                            //             MainAxisAlignment.spaceBetween,
-                            //         children: [
-                            //           Column(
-                            //             crossAxisAlignment:
-                            //                 CrossAxisAlignment.start,
-                            //             children: <Widget>[
-                            //               Text(
-                            //                 "Invite ",
-                            //                 textScaleFactor: mediaQueryData
-                            //                     .textScaleFactor
-                            //                     .clamp(0.5, 1.5)
-                            //                     .toDouble(),
-                            //                 style: TextStyle(
-                            //                     //  color: Colors.white,
-                            //                     fontWeight: FontWeight.w700,
-                            //                     fontSize: SizeConfig
-                            //                             .safeBlockHorizontal *
-                            //                         4),
-                            //               ),
-                            //               SizedBox(height: 5),
-                            //               Text(
-                            //                 "Invite friends and earn rewards",
-                            //                 textScaleFactor: mediaQueryData
-                            //                     .textScaleFactor
-                            //                     .clamp(0.5, 0.8)
-                            //                     .toDouble(),
-                            //                 style: TextStyle(
-                            //                     //       color: Colors.white70,
-                            //                     fontWeight: FontWeight.w300,
-                            //                     fontSize: SizeConfig
-                            //                             .safeBlockHorizontal *
-                            //                         3.4),
-                            //               )
-                            //             ],
-                            //           ),
-                            //           Icon(Icons.arrow_forward_ios_rounded,
-                            //               size: 15)
-                            //         ],
-                            //       ),
-                            //     ),
-                            //   ),
-                            // ),
-                            // ReferralDashboard(),
-                            // Padding(
-                            //   padding: const EdgeInsets.all(15.0),
-                            //   child: InkWell(
-                            //     onTap: () {
-                            //       Navigator.push(
-                            //           context,
-                            //           CupertinoPageRoute(
-                            //               builder: (context) => Rewards()));
-                            //     },
-                            //     child: Container(
-                            //       child: Row(
-                            //         mainAxisAlignment:
-                            //             MainAxisAlignment.spaceBetween,
-                            //         children: [
-                            //           Column(
-                            //             crossAxisAlignment:
-                            //                 CrossAxisAlignment.start,
-                            //             children: <Widget>[
-                            //               Text(
-                            //                 "Your rewards",
-                            //                 textScaleFactor: mediaQueryData
-                            //                     .textScaleFactor
-                            //                     .clamp(0.5, 1.5)
-                            //                     .toDouble(),
-                            //                 style: TextStyle(
-                            //                     //  color: Colors.white,
-                            //                     fontWeight: FontWeight.w700,
-                            //                     fontSize: SizeConfig
-                            //                             .safeBlockHorizontal *
-                            //                         4),
-                            //               ),
-                            //               SizedBox(height: 5),
-                            //               Text(
-                            //                 "Check your rewards",
-                            //                 textScaleFactor: mediaQueryData
-                            //                     .textScaleFactor
-                            //                     .clamp(0.5, 0.8)
-                            //                     .toDouble(),
-                            //                 style: TextStyle(
-                            //                     //       color: Colors.white70,
-                            //                     fontWeight: FontWeight.w300,
-                            //                     fontSize: SizeConfig
-                            //                             .safeBlockHorizontal *
-                            //                         3.4),
-                            //               )
-                            //             ],
-                            //           ),
-                            //           Icon(Icons.arrow_forward_ios_rounded,
-                            //               size: 15)
-                            //         ],
-                            //       ),
-                            //     ),
-                            //   ),
-                            // ),
-                            Padding(
-                              padding: const EdgeInsets.all(15.0),
-                              child: InkWell(
-                                onTap: () {
-                                  showBarModalBottomSheet(
-                                      context: context,
-                                      builder: (context) {
-                                        return Container(
-                                          height: MediaQuery.of(context)
-                                              .size
-                                              .height,
-                                          // child: InAppWebView(
-                                          //     gestureRecognizers:
-                                          //         gestureRecognizers,
-                                          //     initialFile:
-                                          //         'https://wallet.hive.blog/@${prefs.getString('HiveUserName')}'),
-                                          child: WebView(
-                                            gestureRecognizers: Set()
-                                              ..add(
-                                                Factory<
-                                                    VerticalDragGestureRecognizer>(
-                                                  () =>
-                                                      VerticalDragGestureRecognizer(),
-                                                ), // or null
-                                              ),
-                                            gestureNavigationEnabled: true,
-                                            javascriptMode:
-                                                JavascriptMode.unrestricted,
-                                            initialUrl:
-                                                'https://wallet.hive.blog/@${prefs.getString('HiveUserName')}',
-                                          ),
-                                        );
-                                      });
-                                },
-                                child: Container(
-                                  child: Row(
-                                    mainAxisAlignment:
-                                        MainAxisAlignment.spaceBetween,
-                                    children: [
-                                      Column(
-                                        crossAxisAlignment:
-                                            CrossAxisAlignment.start,
-                                        children: <Widget>[
-                                          Text(
-                                            "Your wallet",
-                                            textScaleFactor: mediaQueryData
-                                                .textScaleFactor
-                                                .clamp(0.5, 1.5)
-                                                .toDouble(),
-                                            style: TextStyle(
-                                                //  color: Colors.white,
-                                                fontWeight: FontWeight.w700,
-                                                fontSize: SizeConfig
-                                                        .safeBlockHorizontal *
-                                                    4),
-                                          ),
-                                          SizedBox(height: 5),
-                                          Text(
-                                            "Hive wallet",
-                                            textScaleFactor: mediaQueryData
-                                                .textScaleFactor
-                                                .clamp(0.5, 0.8)
-                                                .toDouble(),
-                                            style: TextStyle(
-                                                //       color: Colors.white70,
-                                                fontWeight: FontWeight.w300,
-                                                fontSize: SizeConfig
-                                                        .safeBlockHorizontal *
-                                                    3.4),
-                                          )
-                                        ],
                                       ),
-                                      Icon(Icons.arrow_forward_ios_rounded,
-                                          size: 15)
+                                      // Padding(
+                                      //   padding:
+                                      //       const EdgeInsets.fromLTRB(15, 8, 0, 8),
+                                      //   child: Column(
+                                      //     children: [
+                                      //
+                                      //       Container(
+                                      //         decoration: BoxDecoration(
+                                      //             color: Color(0xff1a1a1a),
+                                      //             borderRadius:
+                                      //                 BorderRadius.circular(10)),
+                                      //         width:
+                                      //             MediaQuery.of(context).size.width /
+                                      //                 4.5,
+                                      //         height:
+                                      //             MediaQuery.of(context).size.width /
+                                      //                 4.5,
+                                      //         child: Icon(Icons.add),
+                                      //       ),
+                                      //       Padding(
+                                      //         padding: const EdgeInsets.all(8.0),
+                                      //         child: Text("Add a podcast"),
+                                      //       )
+                                      //     ],
+                                      //   ),
+                                      // ),
                                     ],
                                   ),
                                 ),
+                              ],
+                            ),
+                          ),
+                          // Container(
+                          //   height: MediaQuery.of(context).size.height * 0.26,
+                          //   child: Column(
+                          //     crossAxisAlignment: CrossAxisAlignment.start,
+                          //     mainAxisSize: MainAxisSize.min,
+                          //     children: [
+                          //       Padding(
+                          //         padding: const EdgeInsets.symmetric(
+                          //             horizontal: 15, vertical: 10),
+                          //         child: Text(
+                          //           "Your Live Rooms",
+                          //           textScaleFactor: 1.0,
+                          //           style: TextStyle(
+                          //               fontSize: SizeConfig.safeBlockHorizontal * 5),
+                          //         ),
+                          //       ),
+                          //       Container(
+                          //         height: MediaQuery.of(context).size.height * 0.2,
+                          //         child: ListView(
+                          //           scrollDirection: Axis.horizontal,
+                          //           children: [
+                          //             Padding(
+                          //               padding:
+                          //                   const EdgeInsets.fromLTRB(15, 8, 0, 8),
+                          //               child: Column(
+                          //                 children: [
+                          //                   Container(
+                          //                     decoration: BoxDecoration(
+                          //                         color: themeProvider
+                          //                             .isLightTheme ==
+                          //                             true
+                          //                             ? Color(0xffE8E8E8)
+                          //                             : Color(0xff1a1a1a),
+                          //                         borderRadius:
+                          //                             BorderRadius.circular(10)),
+                          //                     width:
+                          //                         MediaQuery.of(context).size.width /
+                          //                             4.5,
+                          //                     height:
+                          //                         MediaQuery.of(context).size.width /
+                          //                             4.5,
+                          //                     child: Icon(Icons.add),
+                          //                   ),
+                          //                   Padding(
+                          //                     padding: const EdgeInsets.all(8.0),
+                          //                     child: Text("Add a podcast"),
+                          //                   )
+                          //                 ],
+                          //               ),
+                          //             ),
+                          //             Padding(
+                          //               padding:
+                          //                   const EdgeInsets.fromLTRB(15, 8, 0, 8),
+                          //               child: Column(
+                          //                 children: [
+                          //                   Container(
+                          //                     decoration: BoxDecoration(
+                          //                         color: Color(0xff1a1a1a),
+                          //                         borderRadius:
+                          //                             BorderRadius.circular(10)),
+                          //                     width:
+                          //                         MediaQuery.of(context).size.width /
+                          //                             4.5,
+                          //                     height:
+                          //                         MediaQuery.of(context).size.width /
+                          //                             4.5,
+                          //                     child: Icon(Icons.add),
+                          //                   ),
+                          //                   Padding(
+                          //                     padding: const EdgeInsets.all(8.0),
+                          //                     child: Text("Add a podcast"),
+                          //                   )
+                          //                 ],
+                          //               ),
+                          //             ),
+                          //             Padding(
+                          //               padding:
+                          //                   const EdgeInsets.fromLTRB(15, 8, 0, 8),
+                          //               child: Column(
+                          //                 children: [
+                          //                   Container(
+                          //                     decoration: BoxDecoration(
+                          //                         color: Color(0xff1a1a1a),
+                          //                         borderRadius:
+                          //                             BorderRadius.circular(10)),
+                          //                     width:
+                          //                         MediaQuery.of(context).size.width /
+                          //                             4.5,
+                          //                     height:
+                          //                         MediaQuery.of(context).size.width /
+                          //                             4.5,
+                          //                     child: Icon(Icons.add),
+                          //                   ),
+                          //                   Padding(
+                          //                     padding: const EdgeInsets.all(8.0),
+                          //                     child: Text("Add a podcast"),
+                          //                   )
+                          //                 ],
+                          //               ),
+                          //             ),
+                          //             Padding(
+                          //               padding:
+                          //                   const EdgeInsets.fromLTRB(15, 8, 0, 8),
+                          //               child: Column(
+                          //                 children: [
+                          //                   Container(
+                          //                     decoration: BoxDecoration(
+                          //                         color: themeProvider
+                          //                             .isLightTheme ==
+                          //                             true
+                          //                             ? Color(0xffE8E8E8)
+                          //                             : Color(0xff1a1a1a),
+                          //                         borderRadius:
+                          //                             BorderRadius.circular(10)),
+                          //                     width:
+                          //                         MediaQuery.of(context).size.width /
+                          //                             4.5,
+                          //                     height:
+                          //                         MediaQuery.of(context).size.width /
+                          //                             4.5,
+                          //                     child: Icon(Icons.add),
+                          //                   ),
+                          //                   Padding(
+                          //                     padding: const EdgeInsets.all(8.0),
+                          //                     child: Text("Add a podcast"),
+                          //                   )
+                          //                 ],
+                          //               ),
+                          //             ),
+                          //           ],
+                          //         ),
+                          //       ),
+                          //     ],
+                          //   ),
+                          // ),
+                          ListTile(
+                            onTap: () {
+                              Navigator.push(
+                                  context,
+                                  CupertinoPageRoute(
+                                      builder: (context) =>
+                                          ReferralProgram()));
+                            },
+                            title: Text(
+                              "Invite ",
+                              textScaleFactor: mediaQueryData.textScaleFactor
+                                  .clamp(0.5, 1.5)
+                                  .toDouble(),
+                              style: TextStyle(
+                                  //  color: Colors.white,
+                                  fontWeight: FontWeight.w700,
+                                  fontSize:
+                                      SizeConfig.safeBlockHorizontal * 4),
+                            ),
+                            trailing: Icon(Icons.arrow_forward_ios_rounded,
+                                size: 15),
+                            subtitle: Text(
+                              "Invite friends and earn rewards",
+                              textScaleFactor: mediaQueryData.textScaleFactor
+                                  .clamp(0.5, 0.8)
+                                  .toDouble(),
+                              style: TextStyle(
+                                  //       color: Colors.white70,
+                                  fontWeight: FontWeight.w300,
+                                  fontSize:
+                                      SizeConfig.safeBlockHorizontal * 3.4),
+                            ),
+                          ),
+                          ListTile(
+                            onTap: () {
+                              Navigator.push(
+                                  context,
+                                  CupertinoPageRoute(
+                                      builder: (context) => Rewards()));
+                            },
+                            title: Text(
+                              "Your rewards",
+                              textScaleFactor: mediaQueryData.textScaleFactor
+                                  .clamp(0.5, 1.5)
+                                  .toDouble(),
+                              style: TextStyle(
+                                  //  color: Colors.white,
+                                  fontWeight: FontWeight.w700,
+                                  fontSize:
+                                      SizeConfig.safeBlockHorizontal * 4),
+                            ),
+                            trailing: Icon(Icons.arrow_forward_ios_rounded,
+                                size: 15),
+                            subtitle: Text(
+                              "Check your rewards",
+                              textScaleFactor: mediaQueryData.textScaleFactor
+                                  .clamp(0.5, 0.8)
+                                  .toDouble(),
+                              style: TextStyle(
+                                  //       color: Colors.white70,
+                                  fontWeight: FontWeight.w300,
+                                  fontSize:
+                                      SizeConfig.safeBlockHorizontal * 3.4),
+                            ),
+                          ),
+                          ListTile(
+                            onTap: () {
+                              Navigator.push(
+                                  context,
+                                  CupertinoPageRoute(
+                                      builder: (context) => PublicProfile(
+                                          userId:
+                                              prefs.getString("userId"))));
+                            },
+                            title: Text(
+                              "Public Profile",
+                              textScaleFactor: mediaQueryData.textScaleFactor
+                                  .clamp(0.5, 1.5)
+                                  .toDouble(),
+                              style: TextStyle(
+                                  //  color: Colors.white,
+                                  fontWeight: FontWeight.w700,
+                                  fontSize:
+                                      SizeConfig.safeBlockHorizontal * 4),
+                            ),
+                            trailing: Icon(Icons.arrow_forward_ios_rounded,
+                                size: 15),
+                            subtitle: Text(
+                              "See how others see you",
+                              textScaleFactor: mediaQueryData.textScaleFactor
+                                  .clamp(0.5, 0.8)
+                                  .toDouble(),
+                              style: TextStyle(
+                                  //       color: Colors.white70,
+                                  fontWeight: FontWeight.w300,
+                                  fontSize:
+                                      SizeConfig.safeBlockHorizontal * 3.4),
+                            ),
+                          ),
+
+                          // Padding(
+                          //   padding: const EdgeInsets.all(15.0),
+                          //   child: InkWell(
+                          //     onTap: () {
+                          //       Navigator.push(
+                          //           context,
+                          //           CupertinoPageRoute(
+                          //               builder: (context) =>
+                          //                   ReferralProgram()));
+                          //     },
+                          //     child: Container(
+                          //       child: Row(
+                          //         mainAxisAlignment:
+                          //             MainAxisAlignment.spaceBetween,
+                          //         children: [
+                          //           Column(
+                          //             crossAxisAlignment:
+                          //                 CrossAxisAlignment.start,
+                          //             children: <Widget>[
+                          //               Text(
+                          //                 "Invite ",
+                          //                 textScaleFactor: mediaQueryData
+                          //                     .textScaleFactor
+                          //                     .clamp(0.5, 1.5)
+                          //                     .toDouble(),
+                          //                 style: TextStyle(
+                          //                     //  color: Colors.white,
+                          //                     fontWeight: FontWeight.w700,
+                          //                     fontSize: SizeConfig
+                          //                             .safeBlockHorizontal *
+                          //                         4),
+                          //               ),
+                          //               SizedBox(height: 5),
+                          //               Text(
+                          //                 "Invite friends and earn rewards",
+                          //                 textScaleFactor: mediaQueryData
+                          //                     .textScaleFactor
+                          //                     .clamp(0.5, 0.8)
+                          //                     .toDouble(),
+                          //                 style: TextStyle(
+                          //                     //       color: Colors.white70,
+                          //                     fontWeight: FontWeight.w300,
+                          //                     fontSize: SizeConfig
+                          //                             .safeBlockHorizontal *
+                          //                         3.4),
+                          //               )
+                          //             ],
+                          //           ),
+                          //           Icon(Icons.arrow_forward_ios_rounded,
+                          //               size: 15)
+                          //         ],
+                          //       ),
+                          //     ),
+                          //   ),
+                          // ),
+                          // ReferralDashboard(),
+                          // Padding(
+                          //   padding: const EdgeInsets.all(15.0),
+                          //   child: InkWell(
+                          //     onTap: () {
+                          //       Navigator.push(
+                          //           context,
+                          //           CupertinoPageRoute(
+                          //               builder: (context) => Rewards()));
+                          //     },
+                          //     child: Container(
+                          //       child: Row(
+                          //         mainAxisAlignment:
+                          //             MainAxisAlignment.spaceBetween,
+                          //         children: [
+                          //           Column(
+                          //             crossAxisAlignment:
+                          //                 CrossAxisAlignment.start,
+                          //             children: <Widget>[
+                          //               Text(
+                          //                 "Your rewards",
+                          //                 textScaleFactor: mediaQueryData
+                          //                     .textScaleFactor
+                          //                     .clamp(0.5, 1.5)
+                          //                     .toDouble(),
+                          //                 style: TextStyle(
+                          //                     //  color: Colors.white,
+                          //                     fontWeight: FontWeight.w700,
+                          //                     fontSize: SizeConfig
+                          //                             .safeBlockHorizontal *
+                          //                         4),
+                          //               ),
+                          //               SizedBox(height: 5),
+                          //               Text(
+                          //                 "Check your rewards",
+                          //                 textScaleFactor: mediaQueryData
+                          //                     .textScaleFactor
+                          //                     .clamp(0.5, 0.8)
+                          //                     .toDouble(),
+                          //                 style: TextStyle(
+                          //                     //       color: Colors.white70,
+                          //                     fontWeight: FontWeight.w300,
+                          //                     fontSize: SizeConfig
+                          //                             .safeBlockHorizontal *
+                          //                         3.4),
+                          //               )
+                          //             ],
+                          //           ),
+                          //           Icon(Icons.arrow_forward_ios_rounded,
+                          //               size: 15)
+                          //         ],
+                          //       ),
+                          //     ),
+                          //   ),
+                          // ),
+                          Padding(
+                            padding: const EdgeInsets.all(15.0),
+                            child: InkWell(
+                              onTap: () {
+                                showBottomSheet(
+                                    context: context,
+                                    builder: (context) {
+                                      return Container(
+                                        height: MediaQuery.of(context)
+                                            .size
+                                            .height,
+                                        // child: InAppWebView(
+                                        //     gestureRecognizers:
+                                        //         gestureRecognizers,
+                                        //     initialFile:
+                                        //         'https://wallet.hive.blog/@${prefs.getString('HiveUserName')}'),
+                                        child: WebView(
+                                          gestureRecognizers: Set()
+                                            ..add(
+                                              Factory<
+                                                  VerticalDragGestureRecognizer>(
+                                                () =>
+                                                    VerticalDragGestureRecognizer(),
+                                              ), // or null
+                                            ),
+                                          gestureNavigationEnabled: true,
+                                          javascriptMode:
+                                              JavascriptMode.unrestricted,
+                                          initialUrl:
+                                              'https://wallet.hive.blog/@${prefs.getString('HiveUserName')}',
+                                        ),
+                                      );
+                                    });
+                              },
+                              child: Container(
+                                child: Row(
+                                  mainAxisAlignment:
+                                      MainAxisAlignment.spaceBetween,
+                                  children: [
+                                    Column(
+                                      crossAxisAlignment:
+                                          CrossAxisAlignment.start,
+                                      children: <Widget>[
+                                        Text(
+                                          "Your wallet",
+                                          textScaleFactor: mediaQueryData
+                                              .textScaleFactor
+                                              .clamp(0.5, 1.5)
+                                              .toDouble(),
+                                          style: TextStyle(
+                                              //  color: Colors.white,
+                                              fontWeight: FontWeight.w700,
+                                              fontSize: SizeConfig
+                                                      .safeBlockHorizontal *
+                                                  4),
+                                        ),
+                                        SizedBox(height: 5),
+                                        Text(
+                                          "Hive wallet",
+                                          textScaleFactor: mediaQueryData
+                                              .textScaleFactor
+                                              .clamp(0.5, 0.8)
+                                              .toDouble(),
+                                          style: TextStyle(
+                                              //       color: Colors.white70,
+                                              fontWeight: FontWeight.w300,
+                                              fontSize: SizeConfig
+                                                      .safeBlockHorizontal *
+                                                  3.4),
+                                        )
+                                      ],
+                                    ),
+                                    Icon(Icons.arrow_forward_ios_rounded,
+                                        size: 15)
+                                  ],
+                                ),
                               ),
                             ),
-                            // Padding(
-                            //   padding: const EdgeInsets.all(15.0),
-                            //   child: GestureDetector(
-                            //     onTap: () {
-                            //       Navigator.push(context,
-                            //           CupertinoPageRoute(widget: Settings()));
-                            //     },
-                            //     child: Container(
-                            //       child: Row(
-                            //         mainAxisAlignment:
-                            //             MainAxisAlignment.spaceBetween,
-                            //         children: [
-                            //           Column(
-                            //             crossAxisAlignment:
-                            //                 CrossAxisAlignment.start,
-                            //             children: <Widget>[
-                            //               Text(
-                            //                 "Setting",
-                            //                 textScaleFactor: mediaQueryData
-                            //                     .textScaleFactor
-                            //                     .clamp(0.5, 1.5)
-                            //                     .toDouble(),
-                            //                 style: TextStyle(
-                            //                     //  color: Colors.white,
-                            //                     fontWeight: FontWeight.w700,
-                            //                     fontSize: SizeConfig
-                            //                             .safeBlockHorizontal *
-                            //                         4),
-                            //               ),
-                            //               SizedBox(height: 5),
-                            //               Text(
-                            //                 "Categories , Languages",
-                            //                 textScaleFactor: mediaQueryData
-                            //                     .textScaleFactor
-                            //                     .clamp(0.5, 0.8)
-                            //                     .toDouble(),
-                            //                 style: TextStyle(
-                            //                     //       color: Colors.white70,
-                            //                     fontWeight: FontWeight.w300,
-                            //                     fontSize: SizeConfig
-                            //                             .safeBlockHorizontal *
-                            //                         3.4),
-                            //               )
-                            //             ],
-                            //           ),
-                            //           Icon(
-                            //             Icons.arrow_forward_ios_rounded,
-                            //             size: 15,
-                            //           )
-                            //         ],
-                            //       ),
-                            //     ),
-                            //   ),
-                            // ),
-                          ]),
-                        ),
+                          ),
+                          // Padding(
+                          //   padding: const EdgeInsets.all(15.0),
+                          //   child: GestureDetector(
+                          //     onTap: () {
+                          //       Navigator.push(context,
+                          //           CupertinoPageRoute(widget: Settings()));
+                          //     },
+                          //     child: Container(
+                          //       child: Row(
+                          //         mainAxisAlignment:
+                          //             MainAxisAlignment.spaceBetween,
+                          //         children: [
+                          //           Column(
+                          //             crossAxisAlignment:
+                          //                 CrossAxisAlignment.start,
+                          //             children: <Widget>[
+                          //               Text(
+                          //                 "Setting",
+                          //                 textScaleFactor: mediaQueryData
+                          //                     .textScaleFactor
+                          //                     .clamp(0.5, 1.5)
+                          //                     .toDouble(),
+                          //                 style: TextStyle(
+                          //                     //  color: Colors.white,
+                          //                     fontWeight: FontWeight.w700,
+                          //                     fontSize: SizeConfig
+                          //                             .safeBlockHorizontal *
+                          //                         4),
+                          //               ),
+                          //               SizedBox(height: 5),
+                          //               Text(
+                          //                 "Categories , Languages",
+                          //                 textScaleFactor: mediaQueryData
+                          //                     .textScaleFactor
+                          //                     .clamp(0.5, 0.8)
+                          //                     .toDouble(),
+                          //                 style: TextStyle(
+                          //                     //       color: Colors.white70,
+                          //                     fontWeight: FontWeight.w300,
+                          //                     fontSize: SizeConfig
+                          //                             .safeBlockHorizontal *
+                          //                         3.4),
+                          //               )
+                          //             ],
+                          //           ),
+                          //           Icon(
+                          //             Icons.arrow_forward_ios_rounded,
+                          //             size: 15,
+                          //           )
+                          //         ],
+                          //       ),
+                          //     ),
+                          //   ),
+                          // ),
+                        ]),
                         // child: Container(
                         //   height: MediaQuery.of(context).size.height,
                         //   child: ListView(children: [
@@ -1925,7 +1922,7 @@ class _ProfileState extends State<Profile> with SingleTickerProviderStateMixin {
                         //                                           if (prefs.getString(
                         //                                                   'HiveUserName') ==
                         //                                               null) {
-                        //                                             showBarModalBottomSheet(
+                        //                                             showBottomSheet(
                         //                                                 context:
                         //                                                     context,
                         //                                                 builder:
@@ -1933,7 +1930,7 @@ class _ProfileState extends State<Profile> with SingleTickerProviderStateMixin {
                         //                                                   return HiveDetails();
                         //                                                 });
                         //                                           } else {
-                        //                                             showBarModalBottomSheet(
+                        //                                             showBottomSheet(
                         //                                                 context:
                         //                                                     context,
                         //                                                 builder:
@@ -2564,7 +2561,7 @@ class _ProfileState extends State<Profile> with SingleTickerProviderStateMixin {
                         //                                                   vertical: 15),
                         //                                           child: InkWell(
                         //                                             onTap: () {
-                        //                                               showBarModalBottomSheet(
+                        //                                               showBottomSheet(
                         //                                                   context:
                         //                                                       context,
                         //                                                   builder:
@@ -2609,7 +2606,7 @@ class _ProfileState extends State<Profile> with SingleTickerProviderStateMixin {
                         //                                                   vertical: 15),
                         //                                           child: InkWell(
                         //                                             onTap: () {
-                        //                                               showBarModalBottomSheet(
+                        //                                               showBottomSheet(
                         //                                                   context:
                         //                                                       context,
                         //                                                   builder:
@@ -2731,7 +2728,7 @@ class _ProfileState extends State<Profile> with SingleTickerProviderStateMixin {
                         //                                                   vertical: 15),
                         //                                           child: InkWell(
                         //                                             onTap: () {
-                        //                                               showBarModalBottomSheet(
+                        //                                               showBottomSheet(
                         //                                                   context:
                         //                                                       context,
                         //                                                   builder:
@@ -2989,17 +2986,17 @@ class _ImportPodcastState extends State<ImportPodcast> {
     });
   }
 
-  void showInSnackBar(String value) {
-    final mediaQueryData = MediaQuery.of(context);
-    _RSSImportKey.currentState.showSnackBar(new SnackBar(
-        backgroundColor: Colors.blue,
-        content: new Text(
-          value,
-          textScaleFactor:
-              mediaQueryData.textScaleFactor.clamp(0.2, 1).toDouble(),
-          style: TextStyle(color: Colors.white),
-        )));
-  }
+  // void showInSnackBar(String value) {
+  //   final mediaQueryData = MediaQuery.of(context);
+  //   _RSSImportKey.currentState.showSnackBar(new SnackBar(
+  //       backgroundColor: Colors.blue,
+  //       content: new Text(
+  //         value,
+  //         textScaleFactor:
+  //             mediaQueryData.textScaleFactor.clamp(0.2, 1).toDouble(),
+  //         style: TextStyle(color: Colors.white),
+  //       )));
+  // }
 
   void verifyOTP() async {
     setState(() {
@@ -3021,7 +3018,7 @@ class _ImportPodcastState extends State<ImportPodcast> {
       setState(() {
         isLoading = false;
       });
-      showInSnackBar('${response.data['msg']}');
+      // showInSnackBar('${response.data['msg']}');
     } else {
       createFromRss();
       setState(() {
@@ -3223,7 +3220,7 @@ class _ImportPodcastState extends State<ImportPodcast> {
           child: IconButton(
             onPressed: () {
               Clipboard.setData(ClipboardData(text: newRSSFeed));
-              showInSnackBar('Copied to Clipboard');
+              // showInSnackBar('Copied to Clipboard');
             },
             icon: Icon(
               Icons.copy,
